Initial commit of full the NEURAL_py package
This is the initial commit of the full NEURAL_py package. NEURAL_py is a python implementation of the MATLAB NEURAL qEEF feature set package (https://github.com/otoolej/qEEG_feature_set/releases/tag/v0.4.1). This commit corresponds to version 0.4.1 of the MATLAB NEURAL package.

A demo file is added as a quick guide how to use the overall functions.