Updated functions to take parameters as input
The previous layout meant that NEURAL_parameters.NEURAL_parameters()
had to be modified if any parameters were to be changed. This may not
have been the best coding practice as the default parameters could be
lost and it made trying out different parameter values difficult.

As it is best to not alter the default parameters, there is now the option
to pass in the users own parameters. This way it makes trying out different
parameter combinations easy without needing to alter the package in any way.

There is a demo example also showing this.